Allow plugins to easily specify validation data
Create a new ValidationData class which holds the method, the mechanism and the
session data to be used when validating the account. This information used to
be read from the provider XML file, but since the session data might contain
application keys which are supposed to be kept secret, we also need a way to
allow these keys to be specified within the account plugin binary.

Now, plugin writers can get the context from the ProviderPluginProcess object
and call setValidationData() on it, before calling
ProviderPluginProcess::exec().