Remove invocation of AccountsUI service
This code is no longer needed: first of all, the name of the accounts-ui
service should not be hardcoded here, because we don't know the name of the
service which is using the library.
In the second place, we want to let control to go to the application chosen by
the user, once the account setup is completed.